requirements.txt
dss-python
altdss
pandas
scipy
matplotlib
